deploy script: put in modules
