+++ deploy/individual/NFTXVault.ts
@@ -16,6 +16,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     log: true,
   });
 
+  // ==
   console.log("Setting new Vault Impl in NFTXVaultFactory...");
   await execute(
     "NFTXVaultFactoryUpgradeableV3",
