--- deploy/individual/TimelockExcludeList.ts
@@ -1,19 +1,0 @@
-import { HardhatRuntimeEnvironment, Network } from "hardhat/types";
-import { DeployFunction } from "hardhat-deploy/types";
-import { utils } from "ethers";
-import deployConfig from "../../deployConfig";
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts, network } = hre;
-  const { deploy, execute } = deployments;
-
-  const { deployer } = await getNamedAccounts();
-  const config = deployConfig[network.name];
-
-  const timelockExcludeList = await deploy("TimelockExcludeList", {
-    from: deployer,
-    log: true,
-  });
-};
-export default func;
-func.tags = ["TimelockExcludeList"];
